refactor: rename legacy package to standard and update related scripts
