+++ cmd/standard/daemon_windows.go
@@ -1,4 +1,4 @@
-package legacy
+package standard
 
 func (d *Daemon) RemoveMemlock() error {
 	// This function is a no-op on Windows.
