--- deploy/legacy/prometheus/network-observability/create-cm.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-kubectl delete cm ama-metrics-prometheus-config-node  -n kube-system 
-kubectl create configmap ama-metrics-prometheus-config-node --from-file=./deploy/legacy/prometheus/cilium/prometheus-config -n kube-system 
-k rollout restart ds ama-metrics-node -n kube-system
