+++ deploy/standard/grafana/dashboards/clusters.json
