+++ deploy/standard/manifests/controller/helm/retina/crds/retina.sh_captures.yaml
