+++ deploy/standard/manifests/controller/helm/retina/crds/retina.sh_metricsconfigurations.yaml
