+++ deploy/standard/manifests/controller/helm/retina/crds/retina.sh_retinaendpoints.yaml
