+++ deploy/standard/manifests/controller/helm/retina/templates/daemonset.yaml
