+++ deploy/standard/manifests/controller/helm/retina/templates/networkobserver.yaml
