+++ deploy/standard/manifests/controller/helm/retina/templates/operator.yaml
