+++ deploy/standard/manifests/controller/helm/retina/templates/servicemonitor.yaml
