+++ deploy/standard/prometheus/ama-metrics-settings-configmap.yaml
