+++ deploy/standard/prometheus/collector-config-template.yml
