+++ deploy/standard/prometheus/deploy-retina-clusters.sh
