+++ deploy/standard/prometheus/network-observability/network-observability-svc.yaml
