+++ deploy/standard/prometheus/retina-windows/prometheus-config
