+++ deploy/standard/prometheus/retina/create-cm.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+kubectl create configmap ama-metrics-prometheus-config-node --from-file=./deploy/standard/prometheus/retina/prometheus-config -n kube-system
