+++ pkg/controllers/operator/capture/suite_test.go
@@ -48,7 +48,7 @@ var _ = BeforeSuite(func() {
 
 	By("Bootstrapping test environment")
 	testEnv = &envtest.Environment{
-		CRDDirectoryPaths:     []string{filepath.Join("..", "..", "..", "..", "deploy", "legacy", "manifests", "controller", "helm", "retina", "crds")},
+		CRDDirectoryPaths:     []string{filepath.Join("..", "..", "..", "..", "deploy", "standard", "manifests", "controller", "helm", "retina", "crds")},
 		ErrorIfCRDPathMissing: true,
 	}
