+++ test/profiles/localctx/values.yaml
@@ -1,6 +1,6 @@
 operator:
   enabled: false
-# Plugins will default to deploy/legacy/manifests/controller/helm/retina/values.yaml
+# Plugins will default to deploy/standard/manifests/controller/helm/retina/values.yaml
 # TODO add all plugins that we want to test here for local context.
 remoteContext: false
 enablePodLevel: true
